requirements.txt
fastapi
pydantic
joblib
nltk
pandas
scikit-learn
sklearn-crfsuite
mlflow
uvicorn